Implement 'draw' for shape classes
This commit adds implementations of SVGNode::draw to the shape classes.
The currently supported styles are 'fill', 'stroke', and 'stroke-width'.
Note: For drawing the stroke of circles and ellipses, 'imagearc' had to
be used instead of 'imageellipse'. This is due to the fact that the
latter ignores the value set by 'imagesetthickness'.