Add "state stack" to SVGRenderingHelper
This commit introduces the concept of a stack of translations and
rendering options to the rendering helper class (SVGRenderingHelper).

The way this works is the following:
1. A parent node calls push() to save the current state.
2. It applies transformations and possibly draws itself.
3. If there are any, it renders its children. Since the transformations
are stored on the renderer instance, they still apply.
4. The parent node then calls pop() to restore the transformations that
were active before the call to push().

That concept allows modifying drawing parameters for a node and its
children while not affecting parents or siblings that are drawn
afterwards, since everything will be restored with pop().