Add SVGRenderer class
This class serves as the abstract base class for the specific renderer
instances to be implemented.
It holds the 'render' method that deals with preparing all the visual
attributes before it calls the implementation with the rendering job's
details.

Subclasses only have to implement 3 abstract methods:
* prepareRenderParams, which takes an array of options and converts them
into a new array of parameters to be passed into the following two
methods. Example tasks for this method include coordinate scaling or
conversion between a radius and width/height.
* renderFill, which uses the parameters to draw the shape's fill.
* renderStroke, which does the same for the stroke.