Apply 'viewBox' clip in rasterizer
Due to the fact that the document size also counts towards the image's
scale, it is possible for the viewBox to end somewhere inside the image
and not at its lower-right corner. The drawing would then overflow and
still be drawn. This behavior is fixed by applying a clip as the final
rasterization step.