Add title type (#47)
* feat: Add title type

The value can be set with setValue as added in #42

* feat(writer): prevent XSS in text nodes

Textnodes are not intended to contain valid XML, this might be
especially problematic if they contain javascript. Hence they need to be
escaped in order to prevent cross site scripting attacks.

* refactor(writer): clearer writeNode method

This should make the writeNode method much easier to follow. It also
prints the text from any SVGNode with value.

* chore: suppress false positive phpmd error

Since a title will not be rasterized at all, it makes sense that we do
not use the provided parameter.

* fix(writer): add missing return

This prevents printing a second tag-closing '/>' which would lead to
invalid xml.