chore(ci): Configure test workflow to also run on PHP 8 (#135)
Adds a job for PHP 8 to the test workflow. Also removes staggered execution
of the workflows, so that all of PHPCS, tests, and coverage can execute
immediately.

Dev dependency meyfa/phpunit-assert-gd is updated from v1.1 to v1.2 for
PHP 8 support.

ImageRenderer and a few test cases are adapted to treat \GdImage on PHP 8
the same as gd resources on earlier PHP versions.