feat: Implement path rendering that respects fill-rule! :)
Fixes #28. The time has finally come that I understand how scanline
algorithms work and was able to implement this by referencing an
existing implementation in the SerenityOS path renderer.

I did a lot of work to adapt that code to PHP and get it to perform as
well as possible. Note that for some images rendering still takes ages,
but that is not due to PathRenderer, rather the BezierApproximator does
not yet take scale into account and oftentimes generates curves with
waaaaay too many segments. I have done benchmarks and this rendering
algorithm is insanely efficient for what it does. If anything can be
optimized, it is probably the looping over and transforming of each
point during preparation.