fix: Round the strokeWidth when outlining shapes in renderers
PHP simply cuts off any decimal part of the stroke width. I have just
come across an example where the stroke width that was expected to be
rendered was 1.99999997 pixels, but of course that meant only 1 pixel
wide was actually rendered. With this change, that problem should be
gone. Until we have antialiasing, it's probably fine to assume any
stroke width of greater than or equal to 1.5px should become 2px.