refactor: Convert lengths to pixels before passing to renderers
This allows for a better separation of concerns between nodes and
renderer implementations. Now, the renderers always get resolved pixel
values for all their position and length inputs. This means they only
have to care about transforming the coordinates into the render target
coordinate space by scaling and offsetting, and not have to care about
resolving strings with units against viewport lengths.

Regarding the last point: The spec clearly states that lengths for
properties such as 'ry', 'height' etc. must be resolved against the
document height instead of width. This is now fixed. Others, such as
'stroke-width' or 'r' need to be resolved against the diagonal.
This also makes this change somewhat mandatory: The 'ellipse' renderer
cannot know whether to resolve both 'rx' and 'ry' against the diagonal
(for rendering a circle) or resolve them against width and height
separately (for rendering an ellipse).