feat: Add SVGRasterizer::pushTransform() and ::popTransform() (#150)
* feat: Add SVGRasterizer::pushTransform() and ::popTransform()

The rasterizer obtains a transform stack, which initially contains just
the viewport scale+offset transform. The new methods for pushing to and
popping from that stack are intended to be used by nodes during render.
If a node has a 'transform' attribute, it can call `pushTransform()`,
apply its own transform on top of the previous one, do its rendering
with that, and finally call `popTransform()` to revert to the previous
transform state so that its siblings and ancestors can render correctly
as well.

I also considered passing the parent transform to each node as an
argument of the `SVGNode::rasterize()` method, but that seemed very
messy because we'd have to remember for every node class to properly
clone the transform before mutating it. With this design, there will
still be code duplication, but it's rather straightforward.

* feat: Make rendering nodes push their transform attribute to the stack

With this change, any 'transform' attribute present on any node or node
container will properly affect the rendering of that node and its
descendants.