refactor: Move path approximation into renderer (#151)
* refactor: Move path approximation into renderer

This improves separation of concern between the path element and the
path renderer. The element should not be concerned with how the commands
get turned into points. By moving this logic into the renderer, we
enable a potential future re-implementation where the path approximator
can take into account the current transform to determine its required
accuracy, which would take care of issue #86, among other things such
as generally improved visual fidelity and performance.

* docs: Correct PathRenderer class documentation

* refactor: Move path parser instantiation into SVGPath

The SVGRasterizer should not be concerned with how path strings are
parsed, but only with providing renderer support. The path syntax is
specific to the SVGPath element, hence instantiation of the parser is
moved into that class.