refactor: Make PathApproximator explicitly stateful
PathApproximator is already keeping some state such as $previousCommand
that is only valid for the currently-processed path. There is the risk
of unexpected results when rendering multiple paths in a row, where the
later paths are affected by state that was not properly reset. By making
the class be explicit about its stateful nature and requiring one
instance per path data, these problems are avoided.