refactor: Make PathApproximator explicitly stateful (#152)
* refactor: Make PathApproximator explicitly stateful

PathApproximator is already keeping some state such as $previousCommand
that is only valid for the currently-processed path. There is the risk
of unexpected results when rendering multiple paths in a row, where the
later paths are affected by state that was not properly reset. By making
the class be explicit about its stateful nature and requiring one
instance per path data, these problems are avoided.

* fix: Add array type to PathApproximator internal methods' arguments

* docs: Leave PathApproximator multiple call behavior unspecified

We don't need support for calling ::approximate() multiple times, so it
seems safer to just not specify what happens when it is called multiple
times. This gives us some leeway in implementation.

* fix: Use list() feature in PathApproximator for more conciseness