feat: Rewrite PathApproximator to take the Transform into account
Fixes #86. Instead of approximating the path first and *then* mapping
the resulting points into image space, they are mapped into image space
first, and then approximated. This means there is exactly the right
amount of visual fidelity, independent of the relationship between path
coordinate space and image coordinate space. If the image is scaled up,
the approximation simply gets more detailed, and if the image is scaled
down, the approximation will also contain fewer points to improve
performance.

This works very nicely for lines and Bézier curves which are invariant
under affine transforms, i.e., there is no downside to changing the
order of operations like this. Yet, I still don't know how to apply the
idea to arcs, which might not be invariant, at least not in SVG's
endpoint representation. Hence, they are still approximated in path
space at the moment. More work is needed here.