fix: Fix bogus path splitting logic in PathApproximator
The logic for deciding when to start and end subpaths was completely
broken, as surfaced by the tests added in this patch. This is fixed by
turning $builder into a class member and adding private methods for
explicitly ending the current subpath and optionally starting a new one
at the current position.