feat: Add guessed scale factor to ArcApproximator
The amount by which the arc segment coordinates are scaled is now
guessed inside PathApproximator::arcTo(), and the result is fed into
ArcApproximator::approximate(). Larger scale factors increase the number
of subdivisions of the arc, and as such the number of points produced.
This should work fine in the most common cases but might degrade in
some circumstances (for extremely skewed coordinate systems perhaps?).