feat: Implement fill-rule support for PolygonRenderer
Fixes #156. This patch moves the path rendering logic into a helper
file, PathRendererImplementation.php. That code can then be called both
from PathRenderer where it was used originally and from PolygonRenderer,
so that polygons and polylines gain fill-rule support as well.

A test is added to check that PolygonRenderer respects fill-rule.