fix: Ensure style properties do not contain surrounding whitespace
Fixes #160. SVGNode::setStyle is modified to trim the input before
storing it, effectively guaranteeing that styles never have leading or
trailing whitespace. Tests are added and/or adapted to look out for this
behavior. Some code could be simplified by removing manual trimming
after retrieval of the style value, since this is no longer needed.