fix: Round pixel coordinates for strokeOpenSubpath
Sometimes, the bezier approximator would return coordinates that are
just ever so slightly smaller than an integer (maybe by a factor of
ten to the power of -13). This meant that imageline() would render the
line an entire pixel further down, resulting in a very un-clean line.
By rounding, we avoid this problem.