docs(readme): Start rewriting the README (fixes #125)
The introductory section is rewritten with better wording. A disclaimer
about PHP-SVG still being in the alpha stage is added.

Installation requirements are now specified.
A link to the old master branch is fixed to point to 'main' instead.

The contributing section is moved to the end of the README and
completely rewritten to be more helpful and welcoming to new
contributors.