docs(readme): Rewrite the README (fixes #125) (#163)
* docs(readme): Start rewriting the README (fixes #125)

The introductory section is rewritten with better wording. A disclaimer
about PHP-SVG still being in the alpha stage is added.

Installation requirements are now specified.
A link to the old master branch is fixed to point to 'main' instead.

The contributing section is moved to the end of the README and
completely rewritten to be more helpful and welcoming to new
contributors.

* chore(license): Mention other contributors in copyright notice

* docs(readme): Add packagist download count badge

Adding a badge to celebrate 1 million plus downloads.

* docs: Merge wiki into README.md

The wiki pages contained additional information that was partly
overlapping with the README. This is now all included directly in the
README so that the wiki can be deleted, and there is no longer any
redundancy.